+++ lib/core/index.ts
@@ -1,2 +1,3 @@
 export * from './log';
 export * from './step';
+export * from './middleware.builder';